Builder api + fix tests
